Rename hockeyapp to appcenter
