+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-* @larusso @alexandershestakov @marcolink 
+* @larusso @alexandershestakov
