+++ .github/SUPPORT.md
@@ -1,5 +1,5 @@
 ## Communicating with other developers
 
 pullrequests for all packages.
-- [GitHub issues](https://github.com/wooga/atlas-wdk-unity/issues): All discussions around bugs, feature requests.
-- [GitHub pull requests](https://github.com/wooga/atlas-wdk-unity/pulls): All discussions bugfixes, important and new additions.
+- [GitHub issues](https://github.com/wooga/atlas-appcenter/issues): All discussions around bugs, feature requests.
+- [GitHub pull requests](https://github.com/wooga/atlas-appcenter/pulls): All discussions bugfixes, important and new additions.
