+++ docs/api/allclasses-frame.html
@@ -32,8 +32,9 @@
     <h1 class="bar">All Classes</h1>
     <div class="indexContainer">
         <ul>
-        <li><a href="wooga/gradle/hockey/HockeyPlugin.html" title="class in wooga.gradle.hockey" target="classFrame">HockeyPlugin</a></li>
-<li><a href="wooga/gradle/hockey/tasks/HockeyUploadTask.html" title="class in wooga.gradle.hockey.tasks" target="classFrame">HockeyUploadTask</a></li>
+        <li><a href="wooga/gradle/appcenter/api/AppCenterBuildInfo.html" title="class in wooga.gradle.appcenter.api" target="classFrame">AppCenterBuildInfo</a></li>
+<li><a href="wooga/gradle/appcenter/AppCenterPlugin.html" title="class in wooga.gradle.appcenter" target="classFrame">AppCenterPlugin</a></li>
+<li><a href="wooga/gradle/appcenter/tasks/AppCenterUploadTask.html" title="class in wooga.gradle.appcenter.tasks" target="classFrame">AppCenterUploadTask</a></li>
 
         </ul>
     </div>
