+++ docs/api/deprecated-list.html
@@ -24,7 +24,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
     <!-- Generated by groovydoc -->
-    <title>Deprecated API (Gradle HockeyApp plugin API)</title>
+    <title>Deprecated API (Gradle AppCenter plugin API)</title>
     
     <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
     <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
@@ -34,7 +34,7 @@
 <body class="center">
 <script type="text/javascript"><!--
 if (location.href.indexOf('is-external=true') == -1) {
-    parent.document.title="Deprecated API (Gradle HockeyApp plugin API)";
+    parent.document.title="Deprecated API (Gradle AppCenter plugin API)";
 }
 //-->
 </script>
@@ -96,7 +96,7 @@ if (location.href.indexOf('is-external=true') == -1) {
         <li><a href="index.html?deprecated-list.html" target="_top">Frames</a></li>
         <li><a href="deprecated-list.html" target="_top">No Frames</a></li>
     </ul>
-    <div class="aboutLanguage"><em>Gradle HockeyApp plugin API</em></div>
+    <div class="aboutLanguage"><em>Gradle AppCenter plugin API</em></div>
     <a name="skip-navbar_bottom">
         <!--   -->
     </a></div>
