+++ docs/api/help-doc.html
@@ -25,7 +25,7 @@
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
         <!-- Generated by groovydoc -->
-        <title>API Help (Gradle HockeyApp plugin API)</title>
+        <title>API Help (Gradle AppCenter plugin API)</title>
         <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
         <link href="groovy.ico" type="image/x-icon" rel="icon">
         
@@ -34,7 +34,7 @@
 <body class="center">
 <script type="text/javascript"><!--
 if (location.href.indexOf('is-external=true') == -1) {
-    parent.document.title="API Help (Gradle HockeyApp plugin API)";
+    parent.document.title="API Help (Gradle AppCenter plugin API)";
 }
 //-->
 </script>
@@ -187,7 +187,7 @@ if (location.href.indexOf('is-external=true') == -1) {
         <li><a href="index.html?help-doc.html" target="_top">Frames</a></li>
         <li><a href="help-doc.html" target="_top">No Frames</a></li>
     </ul>
-    <div class="aboutLanguage"><em>Gradle HockeyApp plugin API</em></div>
+    <div class="aboutLanguage"><em>Gradle AppCenter plugin API</em></div>
     <a name="skip-navbar_bottom">
         <!--   -->
     </a></div>
