+++ docs/api/index.html
@@ -21,7 +21,7 @@
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN" "http://www.w3.org/TR/html4/frameset.dtd">
 <html>
 <head>
-<title>Index (Gradle HockeyApp plugin API)</title>
+<title>Index (Gradle AppCenter plugin API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
 <link href="groovy.ico" type="image/x-icon" rel="icon">
