+++ docs/api/overview-summary.html
@@ -22,7 +22,7 @@
 
 <html>
 <head>
-<title>Overview (Gradle HockeyApp plugin API)</title>
+<title>Overview (Gradle AppCenter plugin API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="keywords" content="Overview">
 <link rel ="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -32,7 +32,7 @@
 <body class="center">
 <script type="text/javascript"><!--
 if (location.href.indexOf('is-external=true') == -1) {
-    parent.document.title="Overview (Gradle HockeyApp plugin API)";
+    parent.document.title="Overview (Gradle AppCenter plugin API)";
 }
 //-->
 </script>
@@ -66,7 +66,7 @@ if (location.href.indexOf('is-external=true') == -1) {
     </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-    <h1 class="title">Gradle HockeyApp plugin API</h1>
+    <h1 class="title">Gradle AppCenter plugin API</h1>
 </div>
 
 <div class="header">
@@ -85,13 +85,19 @@ if (location.href.indexOf('is-external=true') == -1) {
 <tbody>
 
 <tr class="altColor">
-    <td class="colFirst"><a href="wooga/gradle/hockey/package-summary.html">wooga.gradle.hockey</a></td>
+    <td class="colFirst"><a href="wooga/gradle/appcenter/package-summary.html">wooga.gradle.appcenter</a></td>
     <td class="colLast">&nbsp;</td>
 </tr>
 
 
 <tr class="rowColor">
-    <td class="colFirst"><a href="wooga/gradle/hockey/tasks/package-summary.html">wooga.gradle.hockey.tasks</a></td>
+    <td class="colFirst"><a href="wooga/gradle/appcenter/api/package-summary.html">wooga.gradle.appcenter.api</a></td>
+    <td class="colLast">&nbsp;</td>
+</tr>
+
+
+<tr class="altColor">
+    <td class="colFirst"><a href="wooga/gradle/appcenter/tasks/package-summary.html">wooga.gradle.appcenter.tasks</a></td>
     <td class="colLast">&nbsp;</td>
 </tr>
 
@@ -126,7 +132,7 @@ if (location.href.indexOf('is-external=true') == -1) {
         <li><a href="index.html?overview-summary.html" target="_top">Frames</a></li>
         <li><a href="overview-summary.html" target="_top">No Frames</a></li>
     </ul>
-    <div class="aboutLanguage"><em>Gradle HockeyApp plugin API</em></div>
+    <div class="aboutLanguage"><em>Gradle AppCenter plugin API</em></div>
     <a name="skip-navbar_bottom">
         <!--   -->
     </a></div>
