+++ docs/api/wooga/gradle/appcenter/api/package-frame.html
@@ -21,19 +21,19 @@
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-    <title>wooga.gradle.hockey</title>
+    <title>wooga.gradle.appcenter.api</title>
     
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta name="keywords" CONTENT="wooga/gradle/hockey package">
+    <meta name="keywords" CONTENT="wooga/gradle/appcenter/api package">
 
     <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
     <link href="groovy.ico" type="image/x-icon" rel="icon">
-    <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
+    <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
 </head>
 
 <body class="left">
 <div class="indexHeader">
-    <b><a href="package-summary.html" target="classFrame">wooga.gradle.hockey</a></b>
+    <b><a href="package-summary.html" target="classFrame">wooga.gradle.appcenter.api</a></b>
 </div>
 
 
@@ -43,7 +43,7 @@
 <div class="indexContainer">
     <h2>Classes</h2>
     <ul>
-        <li><a href="HockeyPlugin.html" title="class in wooga.gradle.hockey" target="classFrame">HockeyPlugin</a></li>
+        <li><a href="AppCenterBuildInfo.html" title="class in wooga.gradle.appcenter.api" target="classFrame">AppCenterBuildInfo</a></li>
     </ul>
 </div>
