+++ docs/api/wooga/gradle/appcenter/package-frame.html
@@ -21,31 +21,38 @@
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-<title>Overview (Gradle AppCenter plugin API)</title>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="keywords" content="Overview">
+    <title>wooga.gradle.appcenter</title>
+    
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <meta name="keywords" CONTENT="wooga/gradle/appcenter package">
 
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-
+    <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
+    <link href="groovy.ico" type="image/x-icon" rel="icon">
+    <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 </head>
 
 <body class="left">
-<div class="indexHeader"><a href="allclasses-frame.html" target="packageFrame">All Classes</a></div>
+<div class="indexHeader">
+    <b><a href="package-summary.html" target="classFrame">wooga.gradle.appcenter</a></b>
+</div>
+
+
+
+
+
 <div class="indexContainer">
-    <h2 title="Packages">Packages</h2>
-    <ul title="Packages">
-        
-        <li><a href="wooga/gradle/appcenter/package-frame.html"
-               target="packageFrame">wooga.gradle.appcenter</a></li>
-        
-        <li><a href="wooga/gradle/appcenter/api/package-frame.html"
-               target="packageFrame">wooga.gradle.appcenter.api</a></li>
-        
-        <li><a href="wooga/gradle/appcenter/tasks/package-frame.html"
-               target="packageFrame">wooga.gradle.appcenter.tasks</a></li>
-        
+    <h2>Classes</h2>
+    <ul>
+        <li><a href="AppCenterPlugin.html" title="class in wooga.gradle.appcenter" target="classFrame">AppCenterPlugin</a></li>
     </ul>
 </div>
+
+
+
+
+
+
+
+
 </body>
 </html>
-
