+++ docs/api/wooga/gradle/appcenter/package-summary.html
@@ -22,16 +22,16 @@
 <html>
 <head>
 
-<title>wooga.gradle.hockey.tasks (Gradle HockeyApp plugin API)</title>
-<meta name="keywords" content="wooga.gradle.hockey.tasks package">
+<title>wooga.gradle.appcenter (Gradle AppCenter plugin API)</title>
+<meta name="keywords" content="wooga.gradle.appcenter package">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<link rel ="stylesheet" TYPE="text/css" HREF="../../../../stylesheet.css" title="Style">
-<link href="../../../../groovy.ico" type="image/x-icon" rel="shortcut icon">
-<link href="../../../../groovy.ico" type="image/x-icon" rel="icon">
+<link rel ="stylesheet" TYPE="text/css" HREF="../../../stylesheet.css" title="Style">
+<link href="../../../groovy.ico" type="image/x-icon" rel="shortcut icon">
+<link href="../../../groovy.ico" type="image/x-icon" rel="icon">
 <script type="text/javascript">
 function windowTitle()
 {
-    parent.document.title="wooga.gradle.hockey.tasks (Gradle HockeyApp plugin API)";
+    parent.document.title="wooga.gradle.appcenter (Gradle AppCenter plugin API)";
 }
 </script>
 <noscript>
@@ -49,18 +49,18 @@ function windowTitle()
 </a>
 <a name="navbar_top_firstrow"><!-- --></a>
 <ul class="navList" title="Navigation">
-    <li><a href="../../../../overview-summary.html">Overview</a></li>
+    <li><a href="../../../overview-summary.html">Overview</a></li>
     <li class="navBarCell1Rev">Package</li>
     <li>Class</li>
-    <li><a href="../../../../deprecated-list.html">Deprecated</a></li>
-    <li><a href="../../../../index-all.html">Index</a></li>
-    <li><a href="../../../../help-doc.html">Help</a></li>
+    <li><a href="../../../deprecated-list.html">Deprecated</a></li>
+    <li><a href="../../../index-all.html">Index</a></li>
+    <li><a href="../../../help-doc.html">Help</a></li>
 </ul>
 </div>
 
 <div class="subNav">
     <ul class="navList">
-        <li><a href="../../../../index.html?wooga/gradle/hockey/tasks/package-summary.html" target="_top">Frames</a></li>
+        <li><a href="../../../index.html?wooga/gradle/appcenter/package-summary.html" target="_top">Frames</a></li>
         <li><a href="package-summary.html" target="_top">No Frames</a></li>
     </ul>
 </div>
@@ -68,7 +68,7 @@ function windowTitle()
 <!-- ========= END OF TOP NAVBAR ========= -->
 
 <div class="header">
-    <h1 class="title">Package wooga.gradle.hockey.tasks</h1>
+    <h1 class="title">Package wooga.gradle.appcenter</h1>
 </div>
 
 
@@ -90,8 +90,8 @@ function windowTitle()
                     </tr>
 <tr class="altColor">
                             <td class="colOne">
-                                <strong><a href="HockeyUploadTask.html" title="class in wooga/gradle/hockey/tasks">
-                                    HockeyUploadTask
+                                <strong><a href="AppCenterPlugin.html" title="class in wooga/gradle/appcenter">
+                                    AppCenterPlugin
                                 </a></strong>
                             </td>
                             <td></td>
@@ -134,7 +134,7 @@ function windowTitle()
         </ul>
     </div>
 
-    <div class="aboutLanguage"><em>Gradle HockeyApp plugin API</em></div>
+    <div class="aboutLanguage"><em>Gradle AppCenter plugin API</em></div>
     <a name="skip-navbar_bottom">
         <!--   -->
     </a></div>
