+++ docs/api/wooga/gradle/appcenter/tasks/package-frame.html
@@ -21,10 +21,10 @@
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-    <title>wooga.gradle.hockey.tasks</title>
+    <title>wooga.gradle.appcenter.tasks</title>
     
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta name="keywords" CONTENT="wooga/gradle/hockey/tasks package">
+    <meta name="keywords" CONTENT="wooga/gradle/appcenter/tasks package">
 
     <link href="groovy.ico" type="image/x-icon" rel="shortcut icon">
     <link href="groovy.ico" type="image/x-icon" rel="icon">
@@ -33,7 +33,7 @@
 
 <body class="left">
 <div class="indexHeader">
-    <b><a href="package-summary.html" target="classFrame">wooga.gradle.hockey.tasks</a></b>
+    <b><a href="package-summary.html" target="classFrame">wooga.gradle.appcenter.tasks</a></b>
 </div>
 
 
@@ -43,7 +43,7 @@
 <div class="indexContainer">
     <h2>Classes</h2>
     <ul>
-        <li><a href="HockeyUploadTask.html" title="class in wooga.gradle.hockey.tasks" target="classFrame">HockeyUploadTask</a></li>
+        <li><a href="AppCenterUploadTask.html" title="class in wooga.gradle.appcenter.tasks" target="classFrame">AppCenterUploadTask</a></li>
     </ul>
 </div>
