+++ settings.gradle
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-rootProject.name = 'atlas-hockey'
+rootProject.name = 'atlas-appcenter'
