--- src/main/groovy/wooga/gradle/hockey/api/AppCenterBuildInfo.groovy
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019 Wooga GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package wooga.gradle.hockey.api
-
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Optional
-
-class AppCenterBuildInfo {
-    @Input
-    @Optional
-    String branchName
-    @Input
-    @Optional
-    String commitHash
-    @Input
-    @Optional
-    String commitMessage
-}
