+++ src/main/resources/META-INF/gradle-plugins/net.wooga.appcenter.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=wooga.gradle.hockey.HockeyPlugin
+implementation-class=wooga.gradle.appcenter.AppCenterPlugin
