+++ src/test/groovy/wooga/gradle/appcenter/AppCenterPluginActivationSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Wooga GmbH
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,24 +12,15 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package wooga.gradle.hockey.api
+package wooga.gradle.appcenter
 
-interface AppVersion {
-    String getId()
-    String getTitle()
-    String getAppId()
+import nebula.test.PluginProjectSpec
 
-    int getAppSize()
-    int getTimespamp()
-    String getDeviceFamily()
-    String getNotes()
-    String getVersion()
-    String getShortVersion()
-    int getStatus()
-    String getConfigUrl()
-    String getPublicUrl()
-    String getBuildUrl()
+class AppCenterPluginActivationSpec extends PluginProjectSpec {
+    @Override
+    String getPluginName() {
+        return 'net.wooga.appcenter'
+    }
 }
